+++ CHANGELOG.md
@@ -10,10 +10,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - **Icon:** add support for icon sets ([#344](https://github.com/studiometa/ui/pull/344), [090b1b1](https://github.com/studiometa/ui/commit/090b1b1))
 - **TwigExtension:** add a `meta_icon('collection:icon')` Twig function ([#223](https://github.com/studiometa/ui/issues/223), [#344](https://github.com/studiometa/ui/pull/344), [7dfb635](https://github.com/studiometa/ui/commit/7dfb635))
+- Add support for PHP 8.4 ([#348](https://github.com/studiometa/ui/pull/348))
 
 ### Changed
 
 - ⚠️ Remove the atomic folders ([#343](https://github.com/studiometa/ui/pull/343), [a61d8e6](https://github.com/studiometa/ui/commit/a61d8e6))
+- ⚠️ Upgrade twig/twig to ^3.0 ([#348](https://github.com/studiometa/ui/pull/348), [c8e3830](https://github.com/studiometa/ui/commit/c8e3830))
+- ⚠️ Upgrade studiometa/twig-toolkit to ^2.0 ([#348](https://github.com/studiometa/ui/pull/348), [39b43c8](https://github.com/studiometa/ui/commit/39b43c8))
 
 ### Removed